completely refactored AStarSearch class, and made a working test program.
The a* nodes were changed as to use shared pointers, and the overall algorithm was simplified.
The pathfinder implementation was fixed so that it correctly returns the node's successors.
The dummy map was fixed so that it doesn't create segfaults, and now allows for variable size.
It is possible to make the test program output small images of the paths it has calculated by defining the DEBUG flag.